backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
python-multipart>=0.0.9
pydantic>=2.8.2
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.35
asyncpg>=0.30.0
psycopg2-binary>=2.9.10
alembic>=1.13.3
pgvector>=0.3.3
redis>=5.0.8
openai>=1.45.0
anthropic>=0.34.2
langchain>=0.3.20
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=0.2.73
tiktoken>=0.7.0
pandas>=2.2.3
numpy>=2.1.0
httpx>=0.27.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
structlog>=24.4.0
pytest>=8.3.3
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
ruff>=0.6.8
mypy>=1.11.2
click>=8.1.7langgraph>=0.2.0
langchain-core>=0.3.0
pandas>=2.0.0
numpy>=1.26.0
